Run lisp-in-small-pieces.sh as the vagrant user.
Run the lisp-in-small-pieces.sh provisioner script as the vagrant user
and call sudo explicitly when elevated privileges are
required. Otherwise, makepkg will rightly complain and refuse to run as
root.

Also, move $tmpdir out of /tmp and into $HOME, as the /tmp filesystem is
too small to hold all the temporary install files.