Pin the pacman mirrorlist to archive.archlinux.org
This allows the user to install software from an archive of the
official arch repos frozen to the date the vm / docker image were
created. This allows the user to install a text editor, say, without
having to upgrade the entire system, even if a long time has passed
since the image was created.

https://wiki.archlinux.org/index.php/Arch_Linux_Archive